Requeted Changes Updated
1) Added showAlert to a separate file
2) Removed Code Duplicates
3) Corrected Indentation from all codes
Now only YailRuntimeError file is left to remove fatalError from
