+++ appinventor/aicompanionapp/src/ViewController.swift
@@ -111,7 +111,6 @@ public class ViewController: UINavigationController, UITextFieldDelegate {
         let runtime = try! String(contentsOfFile: url)
         interpreter.evalForm(runtime)
         if interpreter.exception != nil {
-          //          fatalError("Unable to initialize runtime: \(interpreter.exception!)")
           let title = "Error"
           let message = "Unable to initialize runtime: \(interpreter.exception!)"
           let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
@@ -120,6 +119,7 @@ public class ViewController: UINavigationController, UITextFieldDelegate {
           if let topController = UIApplication.shared.keyWindow?.rootViewController {
             topController.present(alertController, animated: true, completion: nil)
           }
+          return interpreter
         }
         ViewController._interpreterInitialized = true
       }
