+++ appinventor/components-ios/src/DateTimePickerBase.swift
@@ -30,18 +30,8 @@ class DateTimePickerPadController: PickerPadController, DateTimePickerController
   }
 
   public required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-    let alert = UIAlertController(title: "Error", message: "init(coder:) has not been implemented", preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "OK", style: .default) { _ in
-      // Handle OK action if needed
-    }
-    alert.addAction(okAction)
-    
-    if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-      rootViewController.present(alert, animated: true, completion: nil)
-    }
+    showAlert(message: nil)
     return nil
-    
   }
 
   override func setupViews() {
@@ -107,16 +97,7 @@ class DateTimePickerPhoneController: PickerPhoneController, DateTimePickerContro
   }
 
   public required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-    let alert = UIAlertController(title: "Error", message: "init(coder:) has not been implemented", preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "OK", style: .default) { _ in
-      // Handle OK action if needed
-    }
-    alert.addAction(okAction)
-    
-    if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-      rootViewController.present(alert, animated: true, completion: nil)
-    }
+    showAlert(message: nil)
     return nil
   }
