+++ appinventor/components-ios/src/Map.swift
@@ -1249,18 +1249,8 @@ private class ZoomButton: UIButton {
   }
 
   required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-    let alert = UIAlertController(title: "Error", message: "A Error Occured", preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "OK", style: .default) { _ in
-      // Handle OK action if needed
-    }
-    alert.addAction(okAction)
-    
-    if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-      rootViewController.present(alert, animated: true, completion: nil)
-    }
-    
-    return nil // Return nil to indicate failure
+    showAlert(message: nil)
+    return nil
   }
 
   override var isEnabled: Bool {
