+++ appinventor/components-ios/src/Notifier.swift
@@ -183,10 +183,6 @@ fileprivate class CustomAlertView: UIView {
     super.init(frame: frame)
   }
 
-//  required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-//
-//  }
   required init?(coder aDecoder: NSCoder) {
     super.init(coder: aDecoder)
     showAlert(message: nil)
@@ -484,17 +480,5 @@ open class Notifier: NonvisibleComponent {
 
   @objc open func TextInputCanceled() {
     EventDispatcher.dispatchEvent(of: self, called: "TextInputCanceled")
-  }
-}
-
-func showAlert(message: String?) {
-  if let emessage = message {
-    let alertController = UIAlertController(title: "Error", message: emessage, preferredStyle: .alert)
-    alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
-    UIApplication.shared.keyWindow?.rootViewController?.present(alertController, animated: true, completion: nil)
-  } else {
-    let alertController = UIAlertController(title: "Error", message: "init(coder:) has not been implemented", preferredStyle: .alert)
-    alertController.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
-    UIApplication.shared.keyWindow?.rootViewController?.present(alertController, animated: true, completion: nil)
   }
 }
