+++ appinventor/components-ios/src/Spinner.swift
@@ -29,18 +29,8 @@ fileprivate class SpinnerPadController: PickerPadController, SpinnerController {
   }
 
   public required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-    let alert = UIAlertController(title: "Error", message: "A Error Occured", preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "OK", style: .default) { _ in
-      // Handle OK action if needed
-    }
-    alert.addAction(okAction)
-    
-    if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-      rootViewController.present(alert, animated: true, completion: nil)
-    }
-    
-    return nil // Return nil to indicate failure
+    showAlert(message: nil)
+    return nil
   }
 
 
@@ -89,18 +79,8 @@ fileprivate class SpinnerPhoneController: PickerPhoneController, SpinnerControll
   }
 
   public required init?(coder aDecoder: NSCoder) {
-//    fatalError("init(coder:) has not been implemented")
-    let alert = UIAlertController(title: "Error", message: "A Error Occured", preferredStyle: .alert)
-    let okAction = UIAlertAction(title: "OK", style: .default) { _ in
-      // Handle OK action if needed
-    }
-    alert.addAction(okAction)
-    
-    if let rootViewController = UIApplication.shared.keyWindow?.rootViewController {
-      rootViewController.present(alert, animated: true, completion: nil)
-    }
-    
-    return nil // Return nil to indicate failure
+    showAlert(message: nil)
+    return nil
   }
 
   @objc public func reloadComponents(){
